Add files via upload
Hey! This is my final year Civil Engineering project which is about the evaluation of methods to calculate the evaporation(mm) using different soft computing techniques. 
The softwares used for this mainly is MS Excel, XLSTAT and Orange. The Trend analysis and Sensitivity analysis was carried out for various climate variables.The performance of the best fit method was evaluated using Orange software.